refactored to move some code out of main.py
